refactor AddParticipantTeam - remove widgets to own file to prevent repeated code
